Separation of data generation and seeding for mongodb
